Put testUtils in each of the packages requiring it
- `yarn test` works from the command line
- Jest tests in IntelliJ do **not** work
- still need to exclude testUtils from the package build. The exclude in `tsconfig.build.json` is still not working
