+++ packages/sdk-codegen/src/python.gen.spec.ts
@@ -24,7 +24,7 @@
 
  */
 
-import { TestConfig } from 'packages/sdk-codegen-scripts/src/testUtils'
+import { TestConfig } from './testUtils'
 import { PythonGen } from './python.gen'
 
 const config = TestConfig()
