+++ packages/sdk-codegen/src/typescript.gen.spec.ts
@@ -24,7 +24,7 @@
 
  */
 
-import { TestConfig } from 'packages/sdk-codegen-scripts/src/testUtils'
+import { TestConfig } from './testUtils'
 import { TypescriptGen } from './typescript.gen'
 
 const config = TestConfig()
