+++ packages/sdk/src/rtl/browserServices.spec.ts
@@ -24,7 +24,7 @@
 
  */
 
-import { MockCrypto } from 'packages/sdk-codegen-scripts/src/testUtils'
+import { MockCrypto } from './oauthSession.spec'
 import { BrowserServices } from './browserServices'
 import { DefaultSettings } from './apiSettings'
 import { BrowserCryptoHash, BrowserTransport } from './browserTransport'
