+++ packages/sdk/src/rtl/nodeSettings.spec.ts
@@ -24,7 +24,7 @@
 
  */
 
-import { TestConfig } from 'packages/sdk-codegen-scripts/src/testUtils'
+import { TestConfig } from '../testUtils'
 import { ApiConfig, NodeSettings, NodeSettingsIniFile } from './nodeSettings'
 import { defaultTimeout } from './transport'
 import { boolDefault } from './constants'
