+++ packages/sdk/src/rtl/transport.spec.ts
@@ -24,7 +24,7 @@
 
  */
 
-import { TestConfig } from 'packages/sdk-codegen-scripts/src/testUtils'
+import { TestConfig } from '../testUtils'
 import { encodeParam, ResponseMode, responseMode } from './transport'
 
 const config = TestConfig()
