+++ packages/sdk/src/testUtils/index.ts
@@ -24,5 +24,4 @@
 
  */
 
-export * from './mocks'
 export * from './testUtils'
